# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dokter/pemeriksaan": {
            "get": {
                "tags": [
                    "Pemeriksaan Dokter"
                ],
                "summary": "Ambil semua data pemeriksaan",
                "operationId": "2bd741eae4d8aff769da8877701c72b6",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pemeriksaan Dokter"
                ],
                "summary": "Simpan data pemeriksaan dokter",
                "description": "API untuk mengelola pemeriksaan dokter",
                "operationId": "45fe818951b4f7f2c05e9febb7ee3c26",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ID_PEMERIKSAAN",
                                    "ID_DOKTER",
                                    "ID_PERAWAT",
                                    "ID_RESEP",
                                    "RM",
                                    "DIAGNOSA",
                                    "TINDAKAN",
                                    "STATUS"
                                ],
                                "properties": {
                                    "ID_PEMERIKSAAN": {
                                        "type": "string",
                                        "example": "P001"
                                    },
                                    "ID_DOKTER": {
                                        "type": "string",
                                        "example": "D001"
                                    },
                                    "ID_PERAWAT": {
                                        "type": "string",
                                        "example": "P001"
                                    },
                                    "ID_RESEP": {
                                        "type": "string",
                                        "example": "R001"
                                    },
                                    "RM": {
                                        "type": "string",
                                        "example": "RM001"
                                    },
                                    "DIAGNOSA": {
                                        "type": "string",
                                        "example": "Flu"
                                    },
                                    "TINDAKAN": {
                                        "type": "string",
                                        "example": "Pemeriksaan Umum"
                                    },
                                    "STATUS": {
                                        "type": "string",
                                        "example": "Selesai"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil disimpan"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                }
            }
        },
        "/api/dokter/pemeriksaan/{id}": {
            "get": {
                "tags": [
                    "Pemeriksaan Dokter"
                ],
                "summary": "Ambil data pemeriksaan berdasarkan ID",
                "operationId": "fb7e7971063afd778dab8d4bbcd21195",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "P001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pemeriksaan Dokter"
                ],
                "summary": "Update data pemeriksaan dokter",
                "operationId": "2ddef434e9125d61afe005bc46c9be19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "P001"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ID_DOKTER",
                                    "DIAGNOSA",
                                    "TINDAKAN",
                                    "STATUS"
                                ],
                                "properties": {
                                    "ID_DOKTER": {
                                        "type": "string",
                                        "example": "D002"
                                    },
                                    "DIAGNOSA": {
                                        "type": "string",
                                        "example": "Demam"
                                    },
                                    "TINDAKAN": {
                                        "type": "string",
                                        "example": "Pemeriksaan lanjutan"
                                    },
                                    "STATUS": {
                                        "type": "string",
                                        "example": "Dalam Proses"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diperbarui"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pemeriksaan Dokter"
                ],
                "summary": "Hapus data pemeriksaan berdasarkan ID",
                "operationId": "da1886aa5da27d02665fc84c41fd0ebc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemeriksaan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "P001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}